refactor(agent-service): wire centralized wire/ws types into consumers
Split types/api.ts into types/wire.ts (backend wire DTOs) and types/ws.ts
(this service's own WebSocket frames), and consume them at the real call
sites instead of the duplicate definitions that previously lived inline:

- server.ts imports WsMessage/WsOutgoingMessage/OperatorResultSummaryWs from
  types/ws (local copies removed).
- api/workflow-api.ts imports Workflow/WorkflowPersistRequest from types/wire.
- api/compile-api.ts, agent/util/context-utils.ts, and agent/texera-agent.ts
  import WorkflowFatalError/WorkflowCompilationResponse from types/wire.

Correct WorkflowFatalError to match the backend proto (workflowruntimestate.proto):
`type` is the FatalErrorType enum name (string), with details/operatorId/
workerId/timestamp optional, replacing the inaccurate `type: { name }` /
all-required shape. Type-level only; consumers read only `.message`, so there
is no runtime behavior change.

tsc --noEmit, prettier --check, and bun test (101 pass / 0 fail) all green.
